requirements.txt
pandas
sqlalchemy
psycopg2-binary
requests
yfinance
python-dotenv
pyyaml
